requirements.txt
appdirs==1.4.4
attrs==23.1.0
autopage==0.5.1
Babel==2.12.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cliff==4.3.0
cmd2==2.4.3
cryptography==41.0.1
debtcollector==2.5.0
decorator==5.1.1
dogpile.cache==1.2.1
exceptiongroup==1.1.1
Flask==2.3.2
flask-expects-json==1.7.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keystoneauth1==5.2.1
MarkupSafe==2.1.3
msgpack==1.0.5
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==1.2.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.8.0
oslo.config==9.1.1
oslo.context==5.1.1
oslo.i18n==6.0.0
oslo.log==5.2.0
oslo.serialization==5.1.1
oslo.utils==6.1.0
packaging==23.1
pbr==5.11.1
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
prettytable==3.8.0
pycparser==2.21
pyinotify==0.9.6
pyparsing==3.1.0
pyperclip==1.8.2
pyrsistent==0.19.3
pytest==7.3.2
python-cinderclient==9.3.0
python-dateutil==2.8.2
python-heatclient==3.3.0
python-keystoneclient==5.1.0
python-magnumclient==4.1.0
python-novaclient==18.3.0
python-openstackclient==6.2.0
python-saharaclient==4.1.0
python-swiftclient==4.3.0
pytz==2023.3
PyJWT==2.8.0
PyYAML==6.0
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
simplejson==3.19.1
six==1.16.0
stevedore==5.1.0
tomli==2.0.1
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
